+++ src/macos.rs
@@ -1,4 +1,4 @@
-use winit::keyboard::{KeyCode, NamedKey, NativeKeyCode, PhysicalKey};
+use winit::keyboard::{KeyCode, NamedKey, PhysicalKey};
 
 pub const ALT_STR: &str = "Option";
 pub const LOGO_STR: &str = "Cmd";
@@ -21,8 +21,8 @@ pub fn os_specific_key_name(key: NamedKey) -> Option<&'static str> {
     }
 }
 
-pub fn physical_key_name(physical_key: PhysicalKey) -> String {
-    match physical_key {
+pub fn try_physical_key_name(physical_key: PhysicalKey) -> Option<String> {
+    let s = match physical_key {
         PhysicalKey::Code(key_code) => match key_code {
             KeyCode::Backquote => "`",
             KeyCode::Backslash => "\\",
@@ -93,16 +93,11 @@ pub fn physical_key_name(physical_key: PhysicalKey) -> String {
 
             KeyCode::Escape => "Esc",
 
-            other => return format!("{:?}", other),
-        }
-        .to_string(),
-
-        PhysicalKey::Unidentified(native_key_code) => match native_key_code {
-            NativeKeyCode::Unidentified => "<unknown>".to_string(),
-            NativeKeyCode::Android(sc) => format!("SC{sc}"),
-            NativeKeyCode::MacOS(sc) => format!("SC{sc}"),
-            NativeKeyCode::Windows(sc) => format!("SC{sc}"),
-            NativeKeyCode::Xkb(sc) => format!("SC{sc}"),
+            _ => return None,
         },
-    }
+
+        PhysicalKey::Unidentified(_) => return None,
+    };
+
+    Some(s.to_string())
 }